Add filter option
This patch adds an option to filter the dependencies to be copied into
the libdir. Possible values are a regular expression which is matched on
the dependency jar file names or a predicate which is applied to the
java.io.File object to check for whether it is to be copied.

The config structure is changed so that all libdir specific options now
reside below the :libdir key in the project defintion. :libdir-path is
kept for backwards compatibility, though.